Fixed up awesome/src/lua
